Refactor firebase to a repository file
